requirements_all.txt
jinja2
tqdm
requests
pandas
regex
datasets
soundfile
librosa
torch==2.6.0  # average
torchaudio==2.6.0  # average
transformers==4.45.0  # average
flash-attn==2.7.2.post1
accelerate>=0.26.0

# minicpm
pillow
vector-quantize-pytorch==1.18.5
vocos==0.1.0

# speech-gpt2
einops

# baichuan
av==12.3.0
fire==0.4.0
ujson==5.10.0
easydict==1.13
diffusers==0.24.0
deepspeed
decord==0.6.0
opencv-python==4.10.0.84
imagesize==1.4.1
cairosvg==2.7.1

# glm4voice
hyperpyyaml
conformer==0.3.2
diffusers==0.27.2
huggingface_hub==0.25.2
lightning==2.2.4
rich==13.7.1
gdown==5.1.0
wget==3.2
matplotlib

# lamma-omni
openai-whisper==20240930
fairseq==0.12.2

# qwen2.5-omni
qwen-omni-utils==0.0.4
# transformers==4.52.3

# kimi
loguru
blobfile
timm
torchdyn==1.0.6
# transformers>=4.48.3

# ============ evaluate ==============
WeTextProcessing==1.0.3
#vllm  # choice
jiwer
funasr
zhon
zhconv
onnxruntime==1.18.1
scipy
--find-links https://csukuangfj.github.io/kaldifeat/cuda.html
kaldifeat==1.25.5  # choice, for dialect classify